+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## vT.B.D
+## v0.4.0 (2016-06-02)
 
 ### Added
 
+   - Added implementation for semaphores
+   - Changed the kernel-implementation to be lock-free
+   - Added UART driver for STM32F4 boards (Nucleo and STM32F4-DISCO)
    - Added support for [NUCLEO-F401RE](http://www.st.com/web/catalog/tools/FM116/SC959/SS1532/LN1847/PF260000)
    - Added support for [NXP Kinetis TWR-K64F120M](http://www.nxp.com/products/sensors/accelerometers/3-axis-accelerometers/kinetis-k64-mcu-tower-system-module:TWR-K64F120M)
    - Added support for [Segger SystemView](https://www.segger.com/systemview.html)